moved plugin test to manual testing
