+++ tests/test_dataset_examples.py
@@ -5,17 +5,14 @@ import re
 import unittest
 from collections import defaultdict
 from csv import DictReader
-from datetime import datetime
 from io import TextIOWrapper
-from os import walk
 from pathlib import Path
 from typing import Dict, List, Union
 from unittest.mock import Mock, call, patch
 
-from ingest_validation_tools.error_report import ErrorDict, ErrorReport, InfoDict
+from ingest_validation_tools.error_report import ErrorDict, ErrorReport
 from ingest_validation_tools.upload import Upload
 from tests.fixtures import (
-    PLUGIN_DIR_MAP,
     SCATACSEQ_BOTH_VERSIONS_VALID,
     SCATACSEQ_HIGHER_VERSION_VALID,
     SCATACSEQ_LOWER_VERSION_VALID,
@@ -224,7 +221,7 @@ def get_online_check_fixtures(schema_name: str, dir_path: str) -> Dict:
     return value
 
 
-def _assaytype_side_effect(path: str, row: Dict, *args, **kwargs):
+def assaytype_side_effect(path: str, row: Dict, *args, **kwargs):
     del args, kwargs
     response_dict = _open_and_read_fixtures_file(path)
     dataset_type = row.get("assay_type") if row.get("assay_type") else row.get("dataset_type")
@@ -281,7 +278,7 @@ class TestDatasetExamples(unittest.TestCase):
                     opts = {}
                 with patch(
                     "ingest_validation_tools.validation_utils.get_assaytype_data",
-                    side_effect=lambda row, ingest_url: _assaytype_side_effect(
+                    side_effect=lambda row, ingest_url: assaytype_side_effect(
                         test_dir, row, ingest_url
                     ),
                 ) as mock_assaytype_data:
@@ -345,7 +342,7 @@ class TestDatasetExamples(unittest.TestCase):
     def prep_offline_upload(test_dir: str, opts: Dict) -> Upload:
         with patch(
             "ingest_validation_tools.validation_utils.get_assaytype_data",
-            side_effect=lambda row, ingest_url: _assaytype_side_effect(test_dir, row, ingest_url),
+            side_effect=lambda row, ingest_url: assaytype_side_effect(test_dir, row, ingest_url),
         ):
             with patch("ingest_validation_tools.upload.Upload.online_checks"):
                 upload = Upload(Path(f"{test_dir}/upload"), **opts)
@@ -356,7 +353,7 @@ class TestDatasetExamples(unittest.TestCase):
     def prep_dir_schema_upload(self, test_dir: str, opts: Dict, patch_data: Dict) -> Upload:
         with patch(
             "ingest_validation_tools.validation_utils.get_assaytype_data",
-            side_effect=lambda row, ingest_url: _assaytype_side_effect(test_dir, row, ingest_url),
+            side_effect=lambda row, ingest_url: assaytype_side_effect(test_dir, row, ingest_url),
         ):
             with patch(
                 "ingest_validation_tools.validation_utils.get_possible_directory_schemas",
@@ -455,29 +452,6 @@ class TestDatasetExamples(unittest.TestCase):
                     .get("Directory schema version")
                 )
                 self.assertEqual(dir_schema_version, None)
-
-    def test_info_reporting(self):
-        with patch(
-            "ingest_validation_tools.validation_utils.get_assaytype_data",
-            side_effect=lambda row, ingest_url: _assaytype_side_effect(test_dir, row, ingest_url),
-        ):
-            with patch("ingest_validation_tools.upload.Upload.online_checks"):
-                fake_now = datetime.now()
-                for test_dir in glob.glob(f"examples/plugin-tests/**"):
-                    upload = Upload(
-                        Path(f"{test_dir}/upload"), **PLUGIN_EXAMPLES_OPTS, verbose=False
-                    )
-                    upload.get_errors()
-                    info = upload.get_info()
-                    if info is None:
-                        raise Exception("Info should not be none.")
-                    assert info.git
-                    assert info.time
-                    info.git = "WILL_CHANGE"
-                    info.time = fake_now
-                    info_dict = PLUGIN_DIR_MAP[Path(test_dir).absolute()]
-                    info_dict.time = fake_now
-                    self.assertEqual(info_dict, info)
 
 
 # if __name__ == "__main__":
